enose-sensor-firmware/serial_test/pyproject.toml
[project]
name = "bme688-serial-test"
version = "0.1.0"
description = "ESP32 BME688 传感器串口测试工具"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
    "pyqt6>=6.6.0",
    "pyqtgraph>=0.13.3",
    "numpy>=1.24.0",
]
 

[project.scripts]
bme688-test = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
 
[tool.hatch.build.targets.wheel]
packages = ["."]